requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.8.1
django-simple-history==3.0.0
environ==1.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
